Correct issues in setup.py.mustache
(1) The name of gRPC's PyPI package is "grpcio", not "grpc". Someone
else beat us to the name "grpc". :-(

(2) "Topic :: Software Development :: APIs" does not appear on
https://pypi.python.org/pypi?:action=list_classifiers and thus is not
allowed to appear among a package's classifiers.

(3) "Development Status :: 3 - Alpha" may have been the development
status of the grpcio package at the time that setup.py.mustache was
written, but it is not necessarily the development status of the
package for which a setup.py file is being generated.